external/webkit/LayoutTests/platform/android/webgl-conformance/dumpAsText/conformance/state/gl-geterror-expected.txt

Test getError.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getError was expected value: INVALID_ENUM : should generate INVALID_ENUM
PASS getError was expected value: INVALID_VALUE : should generate INVALID_VALUE
PASS getError was expected value: INVALID_OPERATION : should generate INVALID_OPERATION

This WebGL implementation looks like it uses the many error method
Check is that at least one of the errors is the first error
PASS err1 == gl.INVALID_VALUE || err2 == gl.INVALID_VALUE || err3 == gl.INVALID_VALUE || err4 == gl.INVALID_VALUE || err5 == gl.INVALID_VALUE || err6 == gl.INVALID_VALUE is true
PASS gl.getError() == gl.NO_ERROR is true

PASS successfullyParsed is true

TEST COMPLETE
